Benchmark `@printf "%d %d\n" i i`, et al.
I'm not really proposing this as the syntax, I just wanted to
see how much overhead creating and then calling a closure was
inducing. Here are the results:

  Version 1:

    for i = 1:n
        printf(f, f"%d %d\n", i, i)
    end

    Time: 110.85915565490723 ms

  Version 2:

    with_output_stream(f, @thunk begin
        for i = 1:n
            f"%d %d\n"(i,i)
        end
    end)

    Time: 53.33685874938965 ms

  Version 3:

    with_output_stream(f, @thunk begin
        for i = 1:n
            @printf "%d %d\n" i i
        end
    end)

    Time: 37.446022033691406 ms

Calling `with_output_stream` induces a *huge* amount of overhead
here --- about half of the time is spend on that. However, the
let optimization is not kicking in here either, so that's also an
issue. The final version is getting pretty close to C performance
so if we get rid of the with_output_stream business and get the
let optimization working and make the generated formatting code
only call write once, then I think we're in business.