Run all benchmark scripts 5 times for reliable results.
This is in addition to the inner five times the scripts run the
benchmarks. This makes the Octave and R benchmarks take an insane
amount of time, but eventually it gets done and the results are
much more stable than before, which depended a fair amount on the
exact run.