@printf, @sprintf macros: new printf-like interface.
From here on in, printf-like functionality is accessed by using
the @printf and @sprintf macros. Examples:

    julia> @printf("%d %f\n",123,pi)
    123 3.141593

    julia> sprint(io->@printf(io,"%d %f\n",123,pi))
    "123 3.141593\n"

    julia> @sprintf("%d %f\n",123,pi)
    "123 3.141593\n"

If the first argument to the @printf macro is a literal string then
it is interpreted as a format specification and the current output
stream is used for printing. Otherwise, the first argument is
evaluated and used as an IO handle for printing and the second
argument must be a literal format string. All following arguments
are evaluated and used as values to be printed.